+++ src/SelectMenu/index.js
@@ -1,3 +1,0 @@
-import SelectMenu from './SelectMenu'
-
-export default SelectMenu